backend/requirements.txt
Django>=5.1.7
channels>=4.0.0
channels_redis>=4.1.0
djangorestframework>=3.15.1
djangorestframework-simplejwt
psycopg2-binary
celery>=5.3.6
redis>=5.0.4
pandas>=2.2.2
openpyxl>=3.1.2
python-dotenv
corsheaders 
gunicorn
django-environ 